[tests] Split server tests #838 (#841)
* refactoring

* more refactoring

* fix coverage

* fix coverage

* split user tests

* db_add test

* delete db

* remove test

* db list

* add db user test

* fixes

* fix coverage

+++ agdb_server/src/routes/db.rs
@@ -2,7 +2,8 @@ pub(crate) mod user;
 
 use crate::db::Database;
 use crate::db::DbPool;
-use crate::server_error::ServerError;
+use crate::error_code::ErrorCode;
+use crate::server_error::ServerResponse;
 use crate::user_id::UserId;
 use axum::extract::State;
 use axum::http::StatusCode;
@@ -59,33 +60,28 @@ impl From<Database> for ServerDatabase {
     request_body = ServerDatabase,
     security(("Token" = [])),
     responses(
-         (status = 201, description = "Database added"),
-         (status = 403, description = "Database already exists"),
-         (status = 461, description = "Invalid database name"),
+         (status = 201, description = "db added"),
+         (status = 401, description = "unauthorized"),
+         (status = 465, description = "db already exists"),
+         (status = 467, description = "db invalid"),
     )
 )]
 pub(crate) async fn add(
     user: UserId,
     State(db_pool): State<DbPool>,
     Json(request): Json<ServerDatabase>,
-) -> Result<StatusCode, ServerError> {
+) -> ServerResponse {
     if db_pool.find_database_id(&request.name).is_ok() {
-        return Ok(StatusCode::FORBIDDEN);
+        return Err(ErrorCode::DbExists.into());
     }
 
-    db_pool
-        .add_database(
-            user.0,
-            Database {
-                db_id: None,
-                name: request.name,
-                db_type: request.db_type.to_string(),
-            },
-        )
-        .map_err(|mut e| {
-            e.status = StatusCode::from_u16(461).unwrap();
-            e
-        })?;
+    let db = Database {
+        db_id: None,
+        name: request.name,
+        db_type: request.db_type.to_string(),
+    };
+
+    db_pool.add_database(user.0, db)?;
 
     Ok(StatusCode::CREATED)
 }
@@ -95,38 +91,40 @@ pub(crate) async fn add(
     request_body = ServerDatabaseName,
     security(("Token" = [])),
     responses(
-         (status = 200, description = "Database deleted"),
-         (status = 403, description = "Database not found for user"),
+         (status = 204, description = "db deleted"),
+         (status = 401, description = "unauthorized"),
+         (status = 403, description = "user must be a db admin"),
+         (status = 466, description = "db not found"),
     )
 )]
 pub(crate) async fn delete(
     user: UserId,
     State(db_pool): State<DbPool>,
     Json(request): Json<ServerDatabaseName>,
-) -> Result<StatusCode, ServerError> {
-    let db = db_pool
-        .find_user_database(user.0, &request.name)
-        .map_err(|mut e| {
-            e.status = StatusCode::FORBIDDEN;
-            e
-        })?;
+) -> ServerResponse {
+    let db = db_pool.find_user_database(user.0, &request.name)?;
+
+    if !db_pool.is_db_admin(user.0, db.db_id.unwrap())? {
+        return Ok(StatusCode::FORBIDDEN);
+    }
 
     db_pool.delete_database(db)?;
 
-    Ok(StatusCode::OK)
+    Ok(StatusCode::NO_CONTENT)
 }
 
 #[utoipa::path(get,
     path = "/api/v1/db/list",
     security(("Token" = [])),
     responses(
-         (status = 200, description = "Ok", body = Vec<ServerDatabase>)
+         (status = 200, description = "ok", body = Vec<ServerDatabase>),
+         (status = 401, description = "unauthorized"),
     )
 )]
 pub(crate) async fn list(
     user: UserId,
     State(db_pool): State<DbPool>,
-) -> Result<(StatusCode, Json<Vec<ServerDatabase>>), ServerError> {
+) -> ServerResponse<(StatusCode, Json<Vec<ServerDatabase>>)> {
     let dbs = db_pool
         .find_user_databases(user.0)?
         .into_iter()
@@ -140,23 +138,23 @@ pub(crate) async fn list(
     request_body = ServerDatabaseName,
     security(("Token" = [])),
     responses(
-         (status = 200, description = "Database removed"),
-         (status = 403, description = "Database not found for user"),
+         (status = 204, description = "db removed"),
+         (status = 401, description = "unauthorized"),
+         (status = 466, description = "db not found"),
     )
 )]
 pub(crate) async fn remove(
     user: UserId,
     State(db_pool): State<DbPool>,
     Json(request): Json<ServerDatabaseName>,
-) -> Result<StatusCode, ServerError> {
-    let db = db_pool
-        .find_user_database(user.0, &request.name)
-        .map_err(|mut e| {
-            e.status = StatusCode::FORBIDDEN;
-            e
-        })?;
+) -> ServerResponse {
+    let db = db_pool.find_user_database(user.0, &request.name)?;
+
+    if !db_pool.is_db_admin(user.0, db.db_id.unwrap())? {
+        return Ok(StatusCode::FORBIDDEN);
+    }
 
     db_pool.remove_database(db)?;
 
-    Ok(StatusCode::OK)
+    Ok(StatusCode::NO_CONTENT)
 }
